EduTutor-AI__personalized-learning-with-generative-ai-and-lms-integration-main/Project Files/requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain_ibm
pinecone-client
python-dotenv
requests
google-auth-oauthlib
google-api-python-client
